change the positive data type from DOUBLE to LONG.

+++ src/main/kotlin/com/dsm/rosa/domain/news/domain/News.kt
@@ -11,7 +11,7 @@ class News(
     val content: String,
 
     @Column(name = "positivity")
-    val positivity: Double,
+    val positivity: Long,
 
     @Column(name = "published_date")
     val publishedDate: LocalDate,
